+++ flecsi/exec/leg/tracer.hh
@@ -1,7 +1,9 @@
 #ifndef FLECSI_LEG_EXEC_TRACER_HH
 #define FLECSI_LEG_EXEC_TRACER_HH
 
-#include "flecsi/run/context.hh"
+#include "flecsi/flog.hh"
+#include "flecsi/util/common.hh"
+
 #include <legion.h>
 
 namespace flecsi::exec {
@@ -13,11 +15,8 @@ struct trace {
 
   inline guard make_guard();
 
-  trace()
-    : trace(Legion::Runtime::get_runtime()->generate_dynamic_trace_id()) {}
-  explicit trace(id_t id) : id_(id), skip_(false) {}
-
-  trace(trace &&) = default;
+  trace() : id_(Legion::Runtime::get_runtime()->generate_dynamic_trace_id()) {}
+  [[deprecated("use default constructor")]] explicit trace(id_t id) : id_(id) {}
 
   void skip() {
     skip_ = true;
@@ -31,7 +30,7 @@ private:
       // Call Legion tracing tool
       Legion::Runtime::get_runtime()->begin_trace(
         Legion::Runtime::get_context(),
-        id_,
+        id_.value(),
         false, // logical_only = false
         false, // static_trace  = false
         NULL // std::set<RegionTreeID> *managed = NULL
@@ -43,7 +42,7 @@ private:
   void stop() {
     if(!skip_) {
       Legion::Runtime::get_runtime()->end_trace(
-        Legion::Runtime::get_context(), id_);
+        Legion::Runtime::get_context(), id_.value());
       tracing = false;
     }
     else {
@@ -57,8 +56,8 @@ public:
   }
 
 private:
-  id_t id_;
-  bool skip_;
+  util::move_optional<id_t> id_;
+  bool skip_ = false;
   static inline bool tracing = false;
 
 }; // struct trace