+++ flecsi/exec/mpi/tracer.hh
@@ -11,9 +11,10 @@ struct trace {
   inline guard make_guard();
 
   trace() {}
-  explicit trace(id_t) {}
+  [[deprecated("use default constructor")]] explicit trace(id_t) {}
 
   trace(trace &&) = default;
+  trace & operator=(trace &&) & = default;
 
   void skip() {}