+++ doc/sphinx/src/news.rst
@@ -39,6 +39,10 @@ Deprecated
 
   * ``coloring_slot`` |mdash| use ``specialization::mpi_coloring``
 
+* Execution
+
+  * Constructing a ``trace`` with an id |mdash| use the default constructor
+
 * Topologies
 
   * ``specialization::cslot`` |mdash| use ``mpi_coloring``
@@ -78,6 +82,10 @@ New features
 * Data
 
   * ``topology_slot`` is now movable; empty slots may be detected with ``topology_slot::is_allocated``.
+
+* Execution
+
+  * ``trace`` objects can be move-assigned.
 
 * Topologies